requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlmodel
bcrypt
pyjwt
alembic
pymysql
pydantic-settings
python-dotenv
pytest
httpx
loguru
mysqlclient
pandas
python-multipart